Moved update tables logic to a script and added cli
- It was possible to run .github/scripts/update_synapse_tables.py
  from the command line, but confusing. Now the github action will
  just pass its arguments on to analyze_and_update
- Analyze_and_update_synapse_tables.py can also be called from the
  command line. It allows passing plain table names (with -t flag)
  for convenience. Or -a flag can be used to upload all tables.
